CountFetching: Update counting supporters to use fake json page
The previous method of counting supporters or volunteers wouldn't work.
NationBuilder doesn't allow lists to be dynamically updated and there
is no API endpoint for Filters.  I created a page with an empty
template, except for some hand coded JSON.  This update uses that JSON
page to get the counts.